+++ packages/bloc/test/helpers/complex/complex_bloc.dart
@@ -9,22 +9,29 @@ class ComplexBloc extends Bloc<ComplexEvent, ComplexState> {
   ComplexState get initialState => ComplexStateA();
 
   @override
-  Stream<ComplexState> transform(Stream<ComplexEvent> events, next) {
-    return super.transform(
-      (events as Observable<ComplexEvent>).distinct(),
-      next,
-    );
+  Stream<ComplexState> transformEvents(events, next) {
+    return (events as Observable<ComplexEvent>).switchMap(next);
   }
 
   @override
-  Stream<ComplexState> mapEventToState(ComplexEvent event) {
+  Stream<ComplexState> mapEventToState(ComplexEvent event) async* {
     if (event is ComplexEventA) {
-      return Observable.just(ComplexStateA());
+      yield ComplexStateA();
+    } else if (event is ComplexEventB) {
+      yield ComplexStateB();
+    } else if (event is ComplexEventC) {
+      await Future<void>.delayed(Duration(milliseconds: 100));
+      yield ComplexStateC();
+    } else if (event is ComplexEventD) {
+      await Future<void>.delayed(Duration(milliseconds: 100));
+      yield ComplexStateD();
     }
-    if (event is ComplexEventB) {
-      return Observable.just(ComplexStateB());
-    }
-    return Observable.just(ComplexStateC());
+  }
+
+  @override
+  Stream<ComplexState> transformStates(states) {
+    return (states as Observable<ComplexState>)
+        .debounceTime(Duration(milliseconds: 50));
   }
 
   @override
@@ -36,5 +43,7 @@ class ComplexBloc extends Bloc<ComplexEvent, ComplexState> {
 
   @override
   int get hashCode =>
-      initialState.hashCode ^ mapEventToState.hashCode ^ transform.hashCode;
+      initialState.hashCode ^
+      mapEventToState.hashCode ^
+      transformEvents.hashCode;
 }